adk_core/agents/financial_investment_analysis_agent.py
# Implements the Financial Investment Analysis Agent.
# (Placeholder for future implementation.)

import logging
from google.adk.message import Message # Corrected import

from adk_core.agents.agent import ConstructionAgent
from adk_core.agents.constants import AgentCapabilities, AgentRoles

logger = logging.getLogger(__name__)

class FinancialInvestmentAnalysisAgent(ConstructionAgent):
    """
    Performs financial modeling, investment analysis, and cost-benefit assessments
    to ensure project financial viability and optimal return on investment.
    """
    def __init__(self):
        super().__init__(
            AgentRoles.FINANCIAL_INVESTMENT,
            [
                AgentCapabilities.ANALYSIS,
                AgentCapabilities.FINANCE
            ]
        )
    async def on_message(self, message: Message, context: AgentContext):
        self.context = context
        logger.info(f"Financial Agent received message from '{message.sender_id}': {message.payload}")
        # --- Future Implementation ---
        # This agent would:
        # - Analyze project cost estimates (from Cost/Supply Chain Agent) against budget and funding availability.
        # - Conduct financial simulations (e.g., NPV, ROI, payback period).
        # - Assess investment risks and propose financing strategies.
        # - Generate financial reports and forecasts for stakeholders.
        if message.type == MessageType.REQUEST and message.receiver_id == self.id:
            # Example: A task to "Analyze project ROI" or "Evaluate financing options"
            task = message.payload.get("task")
            logger.info(f"Financial Agent: Received task '{task}'. (Placeholder: No specific action yet).")
            # Implement specific logic here
        elif message.type == MessageType.RESULT:
            logger.info(f"Financial Agent received RESULT from '{message.sender_id}': {message.payload}")
            # This agent would receive cost estimates or budget changes from Project Management/Cost agents.
        else:
            logger.warning(f"Financial Agent: Received unexpected message type '{message.type}' or receiver_id '{message.receiver_id}'.")
